+++ projects/ngx-bidi/CHANGELOG.md
@@ -1,4 +1,6 @@
-## [1.0.5] - 2025-03-04
+## [1.0.6] - 2025-03-04
 
 ### Changed
-Changed name of directive
+Some fixes in README.md
+fixed list of languages in README.md and code
+Added handling direction via dirAuto